// frontend/taskhub-web/src/app/api/generated/model/createWorkItemCommand.ts
/**
 * TaskHub.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Priority } from './priority';


export interface CreateWorkItemCommand { 
    title?: string | null;
    description?: string | null;
    priority?: Priority;
    dueDate?: string | null;
}